add(highlight): Cube highlighting added

+++ core/math/intersect.cpp
@@ -7,7 +7,7 @@ std::optional<double> GetIntersection(Ray ray, AABB box) {
 	double tmax = std::numeric_limits<double>::max();
 	double tmin = std::numeric_limits<double>::min();
 
-	for (int i = 0; i < 3; i++) {
+	for (size_t i = 0; i < 3; i++) {
 		if (ray.direction[i] != 0.f) {
 			double t1 = (box.min[i] - ray.origin[i]) / ray.direction[i];
 			double t2 = (box.max[i] - ray.origin[i]) / ray.direction[i];
